# api/swagger/soccer-manager.swagger.yaml
basePath: /api
definitions:
  ApiResponse:
    properties:
      data: {}
    type: object
...
  CreateTransferRequest:
    properties:
      player_id:
        type: integer
      price:
        type: number
    required:
    - player_id
    - price
    type: object
...
  TransferResponse:
    properties:
      id:
        type: integer
      listed_at:
        type: string
      player_id:
        type: integer
      price:
        type: integer
      seller_team_id:
        type: integer
    type: object
...
  UpdateTransferRequest:
    properties:
      price:
        type: number
    required:
    - price
    type: object
...
  echo.HTTPError:
    properties:
      message: {}
    type: object
host: localhost:80
info:
  contact: {}
  description: Soccer manager api
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Soccer Manager
  version: 1.0.0-alpha0
paths:
...
  /v1/players/{player_id}/transfer:
    get:
      description: Returns a transfer record of a player
      parameters:
      - description: Player ID
        in: path
        name: player_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/TransferResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get transfer of a player
      tags:
      - transfers
...
  /v1/teams/{team_id}/transfers:
    get:
      description: Returns a list of transfers for the provided team ID (paginated)
      parameters:
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: integer
      - description: Pagination cursor
        in: query
        name: cursor
        type: integer
      - description: Number of items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/TransferResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: List transfers by team ID
      tags:
      - transfers
...
  /v1/transfers:
    get:
      description: Returns a list of all transfers (paginated)
      parameters:
      - description: Pagination cursor
        in: query
        name: cursor
        type: integer
      - description: Number of items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/TransferResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: List all transfers
      tags:
      - transfers
    post:
      consumes:
      - application/json
      description: Creates a new transfer listing
      parameters:
      - description: Transfer details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CreateTransferRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/ApiResponse'
            - properties:
                data:
                  type: integer
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - AccessToken: []
      summary: Create transfer
      tags:
      - transfers
  /v1/transfers/{transfer_id}:
    delete:
      description: Deletes a transfer by ID
      parameters:
      - description: Transfer ID
        in: path
        name: transfer_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - AccessToken: []
      summary: Delete transfer
      tags:
      - transfers
    get:
      description: Returns a transfer by id
      parameters:
      - description: Transfer ID
        in: path
        name: transfer_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/TransferResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get transfer with id
      tags:
      - transfers
    put:
      consumes:
      - application/json
      description: Updates an existing transfer by ID
      parameters:
      - description: Transfer ID
        in: path
        name: transfer_id
        required: true
        type: integer
      - description: Updated transfer info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/UpdateTransferRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - AccessToken: []
      summary: Update transfer
      tags:
      - transfers
  /v1/transfers/{transfer_id}/buy:
    post:
      description: Purchases a player in a specific transfer listing
      parameters:
      - description: Transfer ID
        in: path
        name: transfer_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - AccessToken: []
      summary: Buy a player
      tags:
      - transfers
...
schemes:
- http
securityDefinitions:
  AccessToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"